# vulkan: Optimize contiguous copies