Split out from #10206, but the solution I went with is a bit different.

Add a variant of the copy shader for when the tensors are contiguous. Avoid the complex addressing calculations, and do four elements per invocation to hide some other overhead. In #10206, the matrix multiply is much faster if the B matrix is fp16, so there are a lot of these contiguous copies to do that conversion.

Apply similar changes to the scale shader, since scale is always contiguous.

The first commit fixes a bug in `test-backend-ops perf` where it computed the memory footprint of one iteration but then divided by the total time for all iterations.

Before/after on RTX 4070. In the after numbers, the larger copies are more or less framebuffer bandwidth-limited, and the smaller copies are hitting in L2.
```
Before:
  CPY(type_src=f32,type_dst=f16,ne=[256,3072,1,1],permute=[0,0,0,0]):                  80102 runs -    12.86 us/run -     4608 kB/run - 1;34m 341.72 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[512,3072,1,1],permute=[0,0,0,0]):                  43692 runs -    22.91 us/run -     9216 kB/run -  383.71 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[4096,3072,1,1],permute=[0,0,0,0]):                  5016 runs -   206.73 us/run -    73728 kB/run -  340.61 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[16384,16384,1,1],permute=[0,0,0,0]):                 198 runs -  5364.71 us/run -  1572864 kB/run -  288.08 GB/s
  SCALE(type=f32,ne=[256,3072,1,1],scale=2.000000):                    81930 runs -    12.85 us/run -     6144 kB/run -  456.08 GB/s
  SCALE(type=f32,ne=[512,3072,1,1],scale=2.000000):                    43696 runs -    23.07 us/run -    12288 kB/run -  508.00 GB/s
  SCALE(type=f32,ne=[4096,3072,1,1],scale=2.000000):                    4446 runs -   236.93 us/run -    98304 kB/run -  396.26 GB/s
  SCALE(type=f32,ne=[16384,16384,1,1],scale=2.000000):                   204 runs -  4983.04 us/run -  2097152 kB/run -  413.17 GB/s

After:
  CPY(type_src=f32,type_dst=f16,ne=[256,3072,1,1],permute=[0,0,0,0]):                 233024 runs -     4.41 us/run -     4608 kB/run -  997.23 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[512,3072,1,1],permute=[0,0,0,0]):                 160204 runs -     6.34 us/run -     9216 kB/run - 1387.29 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[4096,3072,1,1],permute=[0,0,0,0]):                  6384 runs -   163.27 us/run -    73728 kB/run -  431.28 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[16384,16384,1,1],permute=[0,0,0,0]):                 264 runs -  3908.14 us/run -  1572864 kB/run -  395.45 GB/s
  SCALE(type=f32,ne=[256,3072,1,1],scale=2.000000):                   185708 runs -     5.49 us/run -     6144 kB/run - 1067.64 GB/s
  SCALE(type=f32,ne=[512,3072,1,1],scale=2.000000):                   114702 runs -     8.73 us/run -    12288 kB/run - 1342.44 GB/s
  SCALE(type=f32,ne=[4096,3072,1,1],scale=2.000000):                    4788 runs -   220.35 us/run -    98304 kB/run -  426.08 GB/s
  SCALE(type=f32,ne=[16384,16384,1,1],scale=2.000000):                   204 runs -  5030.21 us/run -  2097152 kB/run -  409.29 GB/s
```


- [x] I have read the [contributing guidelines](https://github.com/ggerganov/llama.cpp/blob/master/CONTRIBUTING.md)
- Self-reported review complexity:
  - [ ] Low
  - [x] Medium
  - [ ] High