@0cc4m 2024-11-12
I did some more benchmarks.

RTX 3090:
```
Before:
  CPY(type_src=f32,type_dst=f16,ne=[256,3072,1,1],permute=[0,0,0,0]):                  94666 runs -    10.66 us/run -     4608 kB/run -  412.15 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[512,3072,1,1],permute=[0,0,0,0]):                  47333 runs -    21.58 us/run -     9216 kB/run -  407.36 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[4096,3072,1,1],permute=[0,0,0,0]):                  6840 runs -   154.11 us/run -    73728 kB/run -  456.91 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[16384,16384,1,1],permute=[0,0,0,0]):                 330 runs -  3222.34 us/run -  1572864 kB/run -  479.61 GB/s

After:
  CPY(type_src=f32,type_dst=f16,ne=[256,3072,1,1],permute=[0,0,0,0]):                 276716 runs -     3.70 us/run -     4608 kB/run - 1188.59 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[512,3072,1,1],permute=[0,0,0,0]):                  80102 runs -    12.60 us/run -     9216 kB/run -  697.57 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[4096,3072,1,1],permute=[0,0,0,0]):                 11400 runs -    91.06 us/run -    73728 kB/run -  773.32 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[16384,16384,1,1],permute=[0,0,0,0]):                 550 runs -  1851.64 us/run -  1572864 kB/run -  834.64 GB/s
```

Tesla P40:
```
Before:
  CPY(type_src=f32,type_dst=f16,ne=[256,3072,1,1],permute=[0,0,0,0]):                  29128 runs -    40.55 us/run -     4608 kB/run -  108.38 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[512,3072,1,1],permute=[0,0,0,0]):                  14564 runs -    75.19 us/run -     9216 kB/run -  116.91 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[4096,3072,1,1],permute=[0,0,0,0]):                  1824 runs -   577.92 us/run -    73728 kB/run -  121.84 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[16384,16384,1,1],permute=[0,0,0,0]):                  88 runs - 12307.07 us/run -  1572864 kB/run -  125.57 GB/s

After:
  CPY(type_src=f32,type_dst=f16,ne=[256,3072,1,1],permute=[0,0,0,0]):                  58256 runs -    18.64 us/run -     4608 kB/run -  235.76 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[512,3072,1,1],permute=[0,0,0,0]):                  29128 runs -    35.05 us/run -     9216 kB/run -  250.77 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[4096,3072,1,1],permute=[0,0,0,0]):                  4104 runs -   255.77 us/run -    73728 kB/run -  275.31 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[16384,16384,1,1],permute=[0,0,0,0]):                 198 runs -  5384.78 us/run -  1572864 kB/run -  287.00 GB/s
```

Radeon Pro VII:
```
Before:
  CPY(type_src=f32,type_dst=f16,ne=[256,3072,1,1],permute=[0,0,0,0]):                  36410 runs -    29.78 us/run -     4608 kB/run -  147.60 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[512,3072,1,1],permute=[0,0,0,0]):                  18205 runs -    67.49 us/run -     9216 kB/run -  130.26 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[4096,3072,1,1],permute=[0,0,0,0]):                  2736 runs -   385.51 us/run -    73728 kB/run -  182.65 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[16384,16384,1,1],permute=[0,0,0,0]):                 132 runs -  7695.48 us/run -  1572864 kB/run -  200.83 GB/s

After:
  CPY(type_src=f32,type_dst=f16,ne=[256,3072,1,1],permute=[0,0,0,0]):                 131076 runs -     7.68 us/run -     4608 kB/run -  571.95 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[512,3072,1,1],permute=[0,0,0,0]):                  76461 runs -    13.51 us/run -     9216 kB/run -  650.84 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[4096,3072,1,1],permute=[0,0,0,0]):                 10944 runs -    91.95 us/run -    73728 kB/run -  765.79 GB/s
  CPY(type_src=f32,type_dst=f16,ne=[16384,16384,1,1],permute=[0,0,0,0]):                 550 runs -  1885.12 us/run -  1572864 kB/run -  819.82 GB/s
```

Looks like a good improvement all around.